+++ src/main/java/org/olegmell/repository/RequestRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface RequestRepository extends CrudRepository<Request, Long> {
+public interface RequestRepository extends CrudRepository<Request, Integer> {
     List<Request> findByTag(String tag);
 }
